Reorganize stories and snapshots to live next to their components
