+++ shared/ui-components/.storybook/config.ts
@@ -1,7 +1,7 @@
 import { configure } from "@storybook/react";
 
 // automatically import all files ending in *.stories.tsx
-const req = require.context("../stories", true, /\.tsx$/);
+const req = require.context("../src", true, /stories\.tsx$/);
 
 function loadStories() {
   req.keys().forEach(req);
