+++ shared/ui-components/__tests__/jest-example.test.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import PageHeader from "../src/headers/page_header";
+import PageHeader from "../src/headers/page_header/page_header";
 import { shallow } from "enzyme";
 
 test("example test", () => {
