+++ shared/ui-components/__tests__/storyshots.test.tsx
@@ -5,7 +5,6 @@ import { createSerializer } from "enzyme-to-json";
 initStoryshots({
   renderer: mount,
   snapshotSerializers: [createSerializer()],
-  integrityOptions: { cwd: __dirname }, // it will start searching from the current directory
   test: multiSnapshotWithOptions({
     renderer: mount
   })
