+++ shared/ui-components/src/atoms/button_example/button_example.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { storiesOf } from "@storybook/react";
-import { Button } from "../src/test-button/button";
+import { Button } from "./button_example";
 
 storiesOf("Button", module)
   .add("with text", () => <Button>Hello Button</Button>)
