+++ shared/ui-components/src/atoms/button_example/button_example.tsx
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import * as React from "react";
 
 export interface IButtonProps {
   children?: React.ReactNode;
