+++ shared/ui-components/src/headers/image_header/image_header.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import ImageCard from "../cards/image_card";
+import ImageCard from "../../cards/image_card";
 
 interface ImageHeaderProps {
   as?: string;
