+++ shared/ui-components/src/headers/page_header/page_header.stories.tsx
@@ -1,0 +1,7 @@
+import * as React from "react";
+import { storiesOf } from "@storybook/react";
+import PageHeader from "./page_header";
+
+storiesOf("Headers|PageHeader", module).add("with text content", () => (
+  <PageHeader>Hello World</PageHeader>
+));
